pyproject.toml
[tool.poetry]
name = "allure-grpc-client"
version = "1.4.2"
description = ""
authors = ["amaym0n <loco.skitls@gmail.com>"]
readme = "README.md"
packages = [
    { include = "grpc_client" }
]

[tool.poetry.dependencies]
grpcio = { version = "*", python = "^3.11" }
grpcio-reflection = { version = "*", python = "^3.11" }
certifi = { version = "*", python = "^3.11" }
allure-pytest = { version = "*", python = "^3.11" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
